+++ src/main/java/com/perscholas/astrodia/models/Port.java
@@ -8,8 +8,7 @@ import lombok.*;
 import javax.persistence.*;
 import java.util.Set;
 
-@Getter
-@Setter
+@Getter @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @RequiredArgsConstructor
@@ -18,7 +17,7 @@ import java.util.Set;
 public class Port {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @NonNull
     private String location;
     @NonNull
