+++ src/main/java/com/perscholas/astrodia/models/Shuttle.java
@@ -8,8 +8,7 @@ import lombok.*;
 import javax.persistence.*;
 import java.util.List;
 
-@Getter
-@Setter
+@Getter @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @RequiredArgsConstructor
@@ -18,7 +17,7 @@ import java.util.List;
 public class Shuttle {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @NonNull
     private String code;
     @NonNull
